--- src/_main/windows.ts
@@ -1,3 +1,0 @@
-import { BrowserWindow } from '../main/windows/window';
-
-export const windows: Assoc<BrowserWindow> = {};
