+++ src/main/config.ts
