+++ src/main/firefox_providers.ts
