+++ src/main/l10n.ts
