+++ src/main/tray/base_template.ts
