+++ src/main/tray/development_template.ts
