+++ src/main/tray/index.ts
