+++ src/main/windows/list.ts
@@ -1,0 +1,3 @@
+import { BrowserWindow } from './window';
+
+export const windows: Assoc<BrowserWindow> = {};
